[TS-485] refactor request in seperate class
